// client/src/app/core/models/autogen/enums.ts
//*** DO NOT MODIFY - CODE IS AUTOGENERATED AND IT WILL BE OVERWRITTEN ***

export enum Equipment {
  Bands = 'bands',
  Barbell = 'barbell',
  BodyOnly = 'bodyonly',
  Cable = 'cable',
  Dumbbell = 'dumbbell',
  EZCurlBar = 'ezcurlbar',
  ExerciseBall = 'exerciseball',
  FoamRoll = 'foamroll',
  Kettlebells = 'kettlebells',
  Machine = 'machine',
  MedicineBall = 'medicineball',
  Other = 'other',
}

export enum Level {
  Beginner = 'beginner',
  Expert = 'expert',
  Intermediate = 'intermediate',
}

export enum Category {
  Cardio = 'cardio',
  OlympicWeightlifting = 'olympicweightlifting',
  Plyometrics = 'plyometrics',
  Powerlifting = 'powerlifting',
  Strength = 'strength',
  Stretching = 'stretching',
  Strongman = 'strongman',
}

export enum Force {
  Pull = 'pull',
  Push = 'push',
  Static = 'static',
}

export enum Mechanic {
  Compound = 'compound',
  Isolation = 'isolation',
}

export enum Muscle {
  Abdominals = 'abdominals',
  Abductors = 'abductors',
  Adductors = 'adductors',
  Biceps = 'biceps',
  Calves = 'calves',
  Chest = 'chest',
  Forearms = 'forearms',
  Glutes = 'glutes',
  Hamstrings = 'hamstrings',
  Lats = 'lats',
  LowerBack = 'lowerback',
  MiddleBack = 'middleback',
  Neck = 'neck',
  Quadriceps = 'quadriceps',
  Shoulders = 'shoulders',
  Traps = 'traps',
  Triceps = 'triceps',
}